Remove loading component and rename member header
