+++ components/ui/report/ReportView.vue
@@ -7,11 +7,9 @@
         :link-stack="breadcrumbsStack"
       />
       <h2>Report details</h2>
-      <LoadingComponent :error="error" :loading="loading">
-        <ReportInfo :report="report" :show-thread="false" :show-message="false" />
-      </LoadingComponent>
+      <ReportInfo :report="report" :show-thread="false" :show-message="false" />
     </section>
-    <section v-if="!loading && !error" class="universal-card">
+    <section class="universal-card">
       <h2>Messages</h2>
       <ConversationThread :thread="thread" :report="report" :update-thread="updateThread" />
     </section>
@@ -21,7 +19,6 @@
 import Breadcrumbs from '~/components/ui/Breadcrumbs.vue'
 import ConversationThread from '~/components/ui/thread/ConversationThread.vue'
 import ReportInfo from '~/components/ui/report/ReportInfo.vue'
-import LoadingComponent from '~/components/ui/LoadingComponent.vue'
 import { addReportMessage } from '~/helpers/threads.js'
 
 const props = defineProps({
@@ -37,44 +34,17 @@ const props = defineProps({
 
 const app = useNuxtApp()
 
-const loading = ref(true)
-const error = ref(null)
-
 const report = ref(null)
 const rawThread = ref(null)
 const thread = computed(() => addReportMessage(rawThread.value, report.value))
 
-onMounted(() => {
-  fetchData()
+await fetchReport().then((result) => {
+  report.value = result
 })
 
-const fetchData = async () => {
-  try {
-    await fetchReport().then((result) => {
-      report.value = result
-    })
-
-    await useBaseFetch(`thread/${report.value.thread_id}`, app.$defaultHeaders()).then((result) => {
-      rawThread.value = result
-    })
-  } catch (err) {
-    onError(err)
-  }
-  loading.value = false
-}
-
-const onError = (err) => {
-  error.value = (
-    err.data ? (err.data.description ? err.data.description : err.data) : err
-  ).toString()
-  app.$notify({
-    group: 'main',
-    title: 'Error loading report',
-    text: error.value,
-    type: 'error',
-  })
-  console.error(err)
-}
+await useBaseFetch(`thread/${report.value.thread_id}`, app.$defaultHeaders()).then((result) => {
+  rawThread.value = result
+})
 
 async function updateThread(newThread) {
   rawThread.value = newThread
