+++ pages/dashboard/notifications.vue
@@ -20,27 +20,25 @@
           </Button>
         </template>
       </div>
-      <LoadingComponent :loading="loading" :error="error">
-        <template v-if="notifications.length > 0">
-          <Chips
-            v-model="selectedType"
-            :items="notifTypes"
-            :format-label="
-              (x) => (x === 'all' ? 'All' : $formatProjectType(x).replace('_', ' ') + 's')
-            "
-            :capitalize="false"
-          />
-          <NotificationItem
-            v-for="notification in notifications"
-            :key="notification.id"
-            v-model:notifications="allNotifs"
-            class="universal-card recessed"
-            :notification="notification"
-            raised
-          />
-        </template>
-        <p v-else>You don't have any unread notifications.</p>
-      </LoadingComponent>
+      <template v-if="notifications.length > 0">
+        <Chips
+          v-model="selectedType"
+          :items="notifTypes"
+          :format-label="
+            (x) => (x === 'all' ? 'All' : $formatProjectType(x).replace('_', ' ') + 's')
+          "
+          :capitalize="false"
+        />
+        <NotificationItem
+          v-for="notification in notifications"
+          :key="notification.id"
+          v-model:notifications="allNotifs"
+          class="universal-card recessed"
+          :notification="notification"
+          raised
+        />
+      </template>
+      <p v-else>You don't have any unread notifications.</p>
     </section>
   </div>
 </template>
@@ -51,7 +49,6 @@ import NotificationItem from '~/components/ui/NotificationItem.vue'
 import Chips from '~/components/ui/Chips.vue'
 import CheckCheckIcon from '~/assets/images/utils/check-check.svg'
 import Breadcrumbs from '~/components/ui/Breadcrumbs.vue'
-import LoadingComponent from '~/components/ui/LoadingComponent.vue'
 
 useHead({
   title: 'Notifications - Modrinth',
@@ -63,9 +60,6 @@ const router = useRouter()
 const history = computed(() => {
   return route.name === 'dashboard-notifications-history'
 })
-
-const loading = ref(true)
-const error = computed(() => (allNotifs.value === null ? 'Failed to load notifications' : null))
 
 const allNotifs = ref(null)
 
@@ -89,33 +83,9 @@ const notifications = computed(() => {
 
 const selectedType = ref('all')
 
-onMounted(() => {
-  fetchData()
+await fetchNotifications().then((result) => {
+  allNotifs.value = result
 })
-
-const fetchData = async () => {
-  try {
-    await fetchNotifications().then((result) => {
-      allNotifs.value = result
-    })
-  } catch (err) {
-    onError(err)
-  }
-  loading.value = false
-}
-
-const onError = (err) => {
-  error.value = (
-    err.data ? (err.data.description ? err.data.description : err.data) : err
-  ).toString()
-  app.$notify({
-    group: 'main',
-    title: 'Error loading notifications',
-    text: error.value,
-    type: 'error',
-  })
-  console.error(err)
-}
 
 function updateRoute() {
   if (history.value) {
