+++ pages/moderation/review.vue
@@ -1,86 +1,84 @@
 <template>
   <section class="universal-card">
     <h2>Review projects</h2>
-    <LoadingComponent :loading="loading" :error="error">
-      <div class="input-group">
-        <Chips
-          v-model="projectType"
-          :items="projectTypes"
-          :format-label="(x) => (x === 'all' ? 'All' : $formatProjectType(x) + 's')"
-        />
-        <button v-if="oldestFirst" class="iconified-button push-right" @click="oldestFirst = false">
-          <SortDescIcon />Sorting by oldest
-        </button>
-        <button v-else class="iconified-button push-right" @click="oldestFirst = true">
-          <SortAscIcon />Sorting by newest
-        </button>
-      </div>
-      <p v-if="projectType !== 'all'" class="project-count">
-        Showing {{ projectsFiltered.length }} {{ projectTypePlural }} of {{ projects.length }} total
-        projects in the queue.
-      </p>
-      <p v-else class="project-count">There are {{ projects.length }} projects in the queue.</p>
-      <p v-if="projectsOver24Hours.length > 0" class="warning project-count">
-        <WarningIcon /> {{ projectsOver24Hours.length }} {{ projectTypePlural }}
-        have been in the queue for over 24 hours.
-      </p>
-      <p v-if="projectsOver48Hours.length > 0" class="danger project-count">
-        <WarningIcon /> {{ projectsOver48Hours.length }} {{ projectTypePlural }}
-        have been in the queue for over 48 hours.
-      </p>
-      <div
-        v-for="project in projectsFiltered.sort((a, b) => {
-          if (oldestFirst) {
-            return b.age - a.age
-          } else {
-            return a.age - b.age
-          }
-        })"
-        :key="`project-${project.id}`"
-        class="universal-card recessed project"
-      >
-        <div class="project-title">
-          <div class="mobile-row">
-            <nuxt-link
-              :to="`/${project.inferred_project_type}/${project.slug}`"
-              class="iconified-stacked-link"
-            >
-              <Avatar :src="project.icon_url" size="xs" no-shadow raised />
-              <span class="stacked">
-                <span class="title">{{ project.title }}</span>
-                <span>{{ $formatProjectType(project.inferred_project_type) }}</span>
-              </span>
-            </nuxt-link>
-          </div>
-          <div class="mobile-row">
-            by
-            <nuxt-link :to="`/user/${project.owner.username}`" class="iconified-link">
-              <Avatar :src="project.owner.avatar_url" circle size="xxs" raised />
-              <span>{{ project.owner.username }}</span>
-            </nuxt-link>
-          </div>
-          <div class="mobile-row">
-            is requesting to be
-            <Badge :type="project.requested_status ? project.requested_status : 'approved'" />
-          </div>
-        </div>
-        <div class="input-group">
+    <div class="input-group">
+      <Chips
+        v-model="projectType"
+        :items="projectTypes"
+        :format-label="(x) => (x === 'all' ? 'All' : $formatProjectType(x) + 's')"
+      />
+      <button v-if="oldestFirst" class="iconified-button push-right" @click="oldestFirst = false">
+        <SortDescIcon />Sorting by oldest
+      </button>
+      <button v-else class="iconified-button push-right" @click="oldestFirst = true">
+        <SortAscIcon />Sorting by newest
+      </button>
+    </div>
+    <p v-if="projectType !== 'all'" class="project-count">
+      Showing {{ projectsFiltered.length }} {{ projectTypePlural }} of {{ projects.length }} total
+      projects in the queue.
+    </p>
+    <p v-else class="project-count">There are {{ projects.length }} projects in the queue.</p>
+    <p v-if="projectsOver24Hours.length > 0" class="warning project-count">
+      <WarningIcon /> {{ projectsOver24Hours.length }} {{ projectTypePlural }}
+      have been in the queue for over 24 hours.
+    </p>
+    <p v-if="projectsOver48Hours.length > 0" class="danger project-count">
+      <WarningIcon /> {{ projectsOver48Hours.length }} {{ projectTypePlural }}
+      have been in the queue for over 48 hours.
+    </p>
+    <div
+      v-for="project in projectsFiltered.sort((a, b) => {
+        if (oldestFirst) {
+          return b.age - a.age
+        } else {
+          return a.age - b.age
+        }
+      })"
+      :key="`project-${project.id}`"
+      class="universal-card recessed project"
+    >
+      <div class="project-title">
+        <div class="mobile-row">
           <nuxt-link
             :to="`/${project.inferred_project_type}/${project.slug}`"
-            class="iconified-button raised-button"
-            ><EyeIcon /> View project</nuxt-link
+            class="iconified-stacked-link"
           >
+            <Avatar :src="project.icon_url" size="xs" no-shadow raised />
+            <span class="stacked">
+              <span class="title">{{ project.title }}</span>
+              <span>{{ $formatProjectType(project.inferred_project_type) }}</span>
+            </span>
+          </nuxt-link>
         </div>
-        <span v-if="project.queued" :class="`submitter-info ${project.age_warning}`">
-          <WarningIcon v-if="project.age_warning" />
-          Submitted
-          <span v-tooltip="$dayjs(project.queued).format('MMMM D, YYYY [at] h:mm A')">{{
-            fromNow(project.queued)
-          }}</span>
-        </span>
-        <span v-else class="submitter-info"><UnknownIcon /> Unknown queue date</span>
+        <div class="mobile-row">
+          by
+          <nuxt-link :to="`/user/${project.owner.username}`" class="iconified-link">
+            <Avatar :src="project.owner.avatar_url" circle size="xxs" raised />
+            <span>{{ project.owner.username }}</span>
+          </nuxt-link>
+        </div>
+        <div class="mobile-row">
+          is requesting to be
+          <Badge :type="project.requested_status ? project.requested_status : 'approved'" />
+        </div>
       </div>
-    </LoadingComponent>
+      <div class="input-group">
+        <nuxt-link
+          :to="`/${project.inferred_project_type}/${project.slug}`"
+          class="iconified-button raised-button"
+          ><EyeIcon /> View project</nuxt-link
+        >
+      </div>
+      <span v-if="project.queued" :class="`submitter-info ${project.age_warning}`">
+        <WarningIcon v-if="project.age_warning" />
+        Submitted
+        <span v-tooltip="$dayjs(project.queued).format('MMMM D, YYYY [at] h:mm A')">{{
+          fromNow(project.queued)
+        }}</span>
+      </span>
+      <span v-else class="submitter-info"><UnknownIcon /> Unknown queue date</span>
+    </div>
   </section>
 </template>
 <script setup>
@@ -93,7 +91,6 @@ import SortDescIcon from '~/assets/images/utils/sort-desc.svg'
 import WarningIcon from '~/assets/images/utils/issues.svg'
 import Badge from '~/components/ui/Badge.vue'
 import { formatProjectType } from '~/plugins/shorthands.js'
-import LoadingComponent from '~/components/ui/LoadingComponent.vue'
 
 useHead({
   title: 'Review projects - Modrinth',
@@ -104,9 +101,6 @@ const app = useNuxtApp()
 const now = app.$dayjs()
 const TIME_24H = 86400000
 const TIME_48H = TIME_24H * 2
-
-const loading = ref(true)
-const error = ref(null)
 
 const projects = ref([])
 const members = ref([])
@@ -146,60 +140,36 @@ const projectTypes = computed(() => {
   return [...set]
 })
 
-onMounted(() => {
-  fetchData()
+await useBaseFetch('moderation/projects?count=1000', app.$defaultHeaders()).then((result) => {
+  projects.value = result
 })
 
-const fetchData = async () => {
-  try {
-    await useBaseFetch('moderation/projects?count=1000', app.$defaultHeaders()).then((result) => {
-      projects.value = result
-    })
+if (projects.value) {
+  const teamIds = projects.value.map((x) => x.team)
 
-    if (projects.value) {
-      const teamIds = projects.value.map((x) => x.team)
+  await useBaseFetch('teams?ids=' + JSON.stringify(teamIds), app.$defaultHeaders()).then(
+    (result) => {
+      members.value = result
 
-      await useBaseFetch('teams?ids=' + JSON.stringify(teamIds), app.$defaultHeaders()).then(
-        (result) => {
-          members.value = result
-
-          projects.value = projects.value.map((project) => {
-            project.owner = members.value
-              .flat()
-              .find((x) => x.team_id === project.team && x.role === 'Owner').user
-            project.age = project.queued ? now - app.$dayjs(project.queued) : Number.MAX_VALUE
-            project.age_warning = ''
-            if (project.age > TIME_24H * 2) {
-              project.age_warning = 'danger'
-            } else if (project.age > TIME_24H) {
-              project.age_warning = 'warning'
-            }
-            project.inferred_project_type = app.$getProjectTypeForUrl(
-              project.project_type,
-              project.loaders
-            )
-            return project
-          })
+      projects.value = projects.value.map((project) => {
+        project.owner = members.value
+          .flat()
+          .find((x) => x.team_id === project.team && x.role === 'Owner').user
+        project.age = project.queued ? now - app.$dayjs(project.queued) : Number.MAX_VALUE
+        project.age_warning = ''
+        if (project.age > TIME_24H * 2) {
+          project.age_warning = 'danger'
+        } else if (project.age > TIME_24H) {
+          project.age_warning = 'warning'
         }
-      )
+        project.inferred_project_type = app.$getProjectTypeForUrl(
+          project.project_type,
+          project.loaders
+        )
+        return project
+      })
     }
-  } catch (err) {
-    onError(err)
-  }
-  loading.value = false
-}
-
-const onError = (err) => {
-  error.value = (
-    err.data ? (err.data.description ? err.data.description : err.data) : err
-  ).toString()
-  app.$notify({
-    group: 'main',
-    title: 'Error loading project queue',
-    text: error.value,
-    type: 'error',
-  })
-  console.error(err)
+  )
 }
 </script>
 <style lang="scss" scoped>
